# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"_36_Enums.Role": {
				"type": "string",
				"enum": [
					"USER",
					"ADMIN"
				]
			},
			"Partial__nom-string--prenom-string--password-string__": {
				"properties": {
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"RegisterBody": {
				"properties": {
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"nom",
					"prenom",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "test",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"properties": {
													"emailVerified": {
														"type": "boolean"
													},
													"role": {
														"$ref": "#/components/schemas/_36_Enums.Role"
													},
													"email": {
														"type": "string"
													},
													"prenom": {
														"type": "string"
													},
													"nom": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"emailVerified",
													"role",
													"email",
													"prenom",
													"nom",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"verifyToken": {
													"type": "string"
												},
												"emailVerified": {
													"type": "boolean"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"password": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"prenom": {
													"type": "string"
												},
												"nom": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"verifyToken",
												"emailVerified",
												"role",
												"password",
												"email",
												"prenom",
												"nom",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"verifyToken": {
													"type": "string"
												},
												"emailVerified": {
													"type": "boolean"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"password": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"prenom": {
													"type": "string"
												},
												"nom": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"verifyToken",
												"emailVerified",
												"role",
												"password",
												"email",
												"prenom",
												"nom",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial__nom-string--prenom-string--password-string__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/profile": {
			"put": {
				"operationId": "UpdateMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"verifyToken": {
													"type": "string"
												},
												"emailVerified": {
													"type": "boolean"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"password": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"prenom": {
													"type": "string"
												},
												"nom": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"verifyToken",
												"emailVerified",
												"role",
												"password",
												"email",
												"prenom",
												"nom",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial__nom-string--prenom-string--password-string__"
							}
						}
					}
				}
			}
		},
		"/private": {
			"get": {
				"operationId": "GetPrivate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Route privée accessible uniquement aux utilisateurs connectés\r\nGET /private\r\nRenvoie un message personnalisé \"Hello {prenom}\"",
				"tags": [
					"Private"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/hello": {
			"get": {
				"operationId": "SayHello",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "un message demandant à vérifier l’adresse mail",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Route d'inscription d'un nouvel utilisateur",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": ": données envoyées par le front (nom, prénom, email, mdp)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterBody",
								"description": ": données envoyées par le front (nom, prénom, email, mdp)"
							}
						}
					}
				}
			}
		},
		"/auth/verify-email": {
			"get": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "un message si l’adresse est bien vérifiée",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Vérifie un email grâce à un token transmis dans l'URL",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"description": ": Token envoyé par email à l’utilisateur",
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "un token JWT (à utiliser pour les routes sécurisées)",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Connecte un utilisateur si email vérifié et mot de passe correct",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": ": email + mot de passe",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginBody",
								"description": ": email + mot de passe"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}